Extract scatterplot and reducers #38
